Added rough design notes for a peer protocol. (#9)
The start of a 'peer' module where the definition of a Peer for this application might live... I added the start of some design notes about a protocol that might be used.

A couple of thoughts that aren't in the those notes. An interesting aspect of this kind of service is that there's really hard guarantees about when anything needs to happen, as long as it does with a reasonable time frame, and reasonable in this case my be measured in hours, maybe even days. That may open up some interesting possibilities for protocol design.